refactor: fix typos and add missing Action suffix to 5 action classes
Renames ApiInvalidFIleTypeException (capital I typo) and adds the
Action suffix to GetDetailLink, GetSite, GetSiteConfig, UpdateSiteConfig,
GetTrackingUserSessions per the {Entity}{Verb}Action convention.
Updates RouteMiddlewareCollector references.
