+++ src/Api/LinkCollection/Action/GetDetailLinkAction.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+namespace LukaLtaApi\Api\LinkCollection\Action;
+
+use LukaLtaApi\Api\ApiAction;
+use LukaLtaApi\Api\LinkCollection\Service\LinkCollectionService;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+
+class GetDetailLinkAction extends ApiAction
+{
+    public function __construct(
+        private readonly LinkCollectionService $service,
+    ) {
+    }
+
+    protected function execute(ServerRequestInterface $request, ResponseInterface $response): ResponseInterface
+    {
+        return $this->service->getDetailLink($request->getAttributes())->getResponse($response);
+    }
+}
