+++ src/Api/WebTracking/Site/Action/GetSiteAction.php
@@ -2,22 +2,22 @@
 
 declare(strict_types=1);
 
-namespace LukaLtaApi\Api\LinkCollection\Action;
+namespace LukaLtaApi\Api\WebTracking\Site\Action;
 
 use LukaLtaApi\Api\ApiAction;
-use LukaLtaApi\Api\LinkCollection\Service\LinkCollectionService;
+use LukaLtaApi\Api\WebTracking\SiteConfig\Service\SiteConfigService;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
-class GetDetailLink extends ApiAction
+class GetSiteAction extends ApiAction
 {
     public function __construct(
-        private readonly LinkCollectionService $service,
+        private readonly SiteConfigService $service,
     ) {
     }
 
     protected function execute(ServerRequestInterface $request, ResponseInterface $response): ResponseInterface
     {
-        return $this->service->getDetailLink($request->getAttributes())->getResponse($response);
+        return $this->service->getSite((int)$request->getAttribute('siteId'))->getResponse($response);
     }
 }
