+++ src/Api/WebTracking/SiteConfig/Action/GetSiteConfigAction.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace LukaLtaApi\Api\WebTracking\Site\Action;
+namespace LukaLtaApi\Api\WebTracking\SiteConfig\Action;
 
 use LukaLtaApi\Api\ApiAction;
 use LukaLtaApi\Api\WebTracking\SiteConfig\Service\SiteConfigService;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
-class GetSite extends ApiAction
+class GetSiteConfigAction extends ApiAction
 {
     public function __construct(
         private readonly SiteConfigService $service,
@@ -18,6 +18,6 @@ class GetSite extends ApiAction
 
     protected function execute(ServerRequestInterface $request, ResponseInterface $response): ResponseInterface
     {
-        return $this->service->getSite((int)$request->getAttribute('siteId'))->getResponse($response);
+        return $this->service->getSiteConfig($request->getAttribute('siteId'))->getResponse($response);
     }
 }
